% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.nearest.R
\name{approx.nearest}
\alias{approx.nearest}
\title{approx.nearest}
\usage{
approx.nearest(x, y, xi)
}
\arguments{
\item{x}{numeric vector giving the coordinates of the points to be interpolated}

\item{y}{corresponding values to x}

\item{xi}{set of numeric values specifying where interpolation has to take place}
}
\value{
zoo-object containing the interpolated data
}
\description{
auxiliary function for applying the function 'InterpolationMethod', approximate a time series using the nearest neighbour
}
\examples{

}
