% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClosestElement.R
\name{ClosestElement}
\alias{ClosestElement}
\title{Get closest element of a vector}
\usage{
ClosestElement(xvector, x, type = "N")
}
\arguments{
\item{xvector}{a vector of values}

\item{x}{the value to find the closest match to}

\item{type}{}
}
\description{
Get closest element of a vector
}
