% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrTimser.R
\name{detrTimser}
\alias{detrTimser}
\title{detrTimser}
\usage{
detrTimser(timeseries, detr = TRUE)
}
\arguments{
\item{timeseries}{time series to be detrendend}

\item{detr}{TRUE (detrending) or FALSE (no detrending)}
}
\value{
zoo-object of the (not) detrendend time series
}
\description{
auxiliary function for applying the functions 'GenNullHypPair' & 'CorQuantilesNullHyp', detrending a time series by removing a linear trend
}
\examples{

}
