% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTimserSlopes.R
\name{estimateTimserSlopes}
\alias{estimateTimserSlopes}
\title{estimateTimserSlopes}
\usage{
estimateTimserSlopes(timeseries1, timeseries2, int.step)
}
\arguments{
\item{timeseries1, timeseries2}{(zoo-object)}

\item{int.step}{equidistant time step of the interpolated irregular time series}
}
\value{
$s1,$s2 fitted slopes of both time series
}
\description{
estimating the slope from a power spectrum of a pair of time series by a linear fit in the frequency range of 
the inverse half of the length of the overlapping time window of both time series and the inverse twofold maximum
mean resolution of both time series
}
\examples{
timeseries1 <- zoo(rnorm(100), order.by=sort(runif(100,min=1,max=1000)))
timeseries2 <- zoo(rnorm(100), order.by=sort(runif(100,min=1,max=1000)))
estimateTimserSlopes(timeseries1,timeseries2,1)
}
