% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.weights.R
\name{gauss.weights}
\alias{gauss.weights}
\title{gauss.weights}
\usage{
gauss.weights(fc)
}
\arguments{
\item{fc}{cut-off frequency}
}
\value{
filter weights
}
\description{
auxiliary function to get the filter weights of the Gaussian filter of a certain cut-off frequency
}
\examples{

}
