% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenNullHypPair.R
\name{GenNullHypPair}
\alias{GenNullHypPair}
\title{GenNullHypPair}
\usage{
GenNullHypPair(n = max(c(max(index(timser1)), max(index(timser2)))),
  beta.noise1, beta.noise2, time1, time2, detr)
}
\arguments{
\item{n}{length of the power-law time series which have to be generated}

\item{beta.noise1, beta.noise2}{slopes from the power spectrum of the time series to be generated}

\item{time1, time2}{time points at which the simulated power-law has to be subsampled}

\item{detr}{TRUE for removing a linear trend, else FALSE}
}
\value{
zoo-objects of the time series
}
\description{
generating a pair of subsampled power-law time series
}
\examples{

}
