% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrateLinearInterpolation.R
\name{IntegrateLinearInterpolation}
\alias{IntegrateLinearInterpolation}
\title{IntegrateLinearInterpolation}
\usage{
IntegrateLinearInterpolation(tx.pos, x.pos)
}
\arguments{
\item{tx.pos}{time points of the time series covered by the filter window and used during integrand interpolation}

\item{x.pos}{product of filter weight and observation at time points of the time series covered by the filter window and used during integrand interpolation}
}
\value{
estimate of the filter integral
}
\description{
auxiliary function for applying the function 'IntegrandInterpolationMethod', approximate solution of the filter integral by linear interpolation
}
\examples{

}
