% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpolationMethod.R
\name{InterpolationMethod}
\alias{InterpolationMethod}
\title{InterpolationMethod}
\usage{
InterpolationMethod(X, fc, dt, timser.length, int.method = c("linear",
  "nearest"), appliedFilter = c("gauss", "runmean", "lowpass"), k = 5)
}
\arguments{
\item{X}{time series (zoo-object)}

\item{fc}{cut-off frequency}

\item{dt}{regular inter-observation time step of the interpolation}

\item{timser.length}{oldest age of the time series}

\item{int.method}{kind of interpolation (linear, nearest neighbor)}

\item{appliedFilter}{time domain filter (gauss, runmean, lowpass)}

\item{k}{scaling factor to define the sharpness of the lowpass}
}
\value{
filtered time series (zoo-object)
}
\description{
linear or nearest neighbor interpolation of the irregular time series before convolution with time domain filter
}
\examples{
timser.irreg <- zoo(rnorm(100), order.by=sort(runif(100,min=1,max=1000)))
timser.irregSmooth <- InterpolationMethod(X=timser.irreg, fc=1/50, dt=10, timser.length=max(index(timser.irreg)), int.method="linear", appliedFilter="lowpass", k=5)
plot(timser.irreg,type="l",col="black")
lines(timser.irregSmooth,col="limegreen",lwd=3)
}
