% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lowpass.R
\name{Lowpass}
\alias{Lowpass}
\title{calculate weights for lowpass filter}
\usage{
Lowpass(omega.c, n = 9, sample = 1, convergence = TRUE)
}
\arguments{
\item{omega.c}{cutoff frequency}

\item{n}{length of the filter, has to be odd}

\item{sample}{sampling rate of the timeseries on which the filter will be applied (1/deltat)}

\item{convergence}{TRUE: smoothed least square lowpass; FALSE = unsmoothed}
}
\value{
vector of filter weights
}
\description{
Derive the (smoothed) least square lowpass, given the cutoff frequency omega.c and the length of the filter n
based on Bloomfield 1976
}
\author{
Thomas Laepple
}
