% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowpass.weights.R
\name{lowpass.weights}
\alias{lowpass.weights}
\title{lowpass.weights}
\usage{
lowpass.weights(fc, k = 5)
}
\arguments{
\item{fc}{cut-off frequency}

\item{k}{scaling factor to define the sharpness of the lowpass}
}
\value{
filter weights
}
\description{
auxiliary function to get the filter weights of the low-pass filter (based on Bloomfield, 1976) of a certain cut-off frequency, necessitates the auxiliary function 'Lowpass'
}
\examples{

}
