% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmean.weights.R
\name{runmean.weights}
\alias{runmean.weights}
\title{runmean.weights}
\usage{
runmean.weights(fc)
}
\arguments{
\item{fc}{cut-off frequency}
}
\value{
filter weights
}
\description{
auxiliary function to get the filter weights of the moving average filter of a certain cut-off frequency
}
\examples{

}
