% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPowerlaw.R
\name{SimPowerlaw}
\alias{SimPowerlaw}
\title{Simulate a random timeseries with a powerlaw spectrum}
\usage{
SimPowerlaw(beta, N)
}
\arguments{
\item{beta}{slope}

\item{N}{length of timeseries to be generated}
}
\value{
vector containing the timeseries
}
\description{
Simulate a random timeseries with a powerlaw spectrum
}
\details{
Method: FFT white noise, rescale, FFT back, the result is scaled to variance 1
}
\author{
Thomas Laepple
}
