% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeFit.R
\name{SlopeFit}
\alias{SlopeFit}
\title{Fit a power-law to the spectrum}
\usage{
SlopeFit(spec, freq.start = NULL, freq.end = NULL, bDebug = TRUE,
  breaks = NULL, indexRemove = NULL, df.log = 0.05, i.fStart = 4)
}
\arguments{
\item{freq.start}{vector containing the start frequencies of the fitting interval(s)}

\item{freq.end}{vector containing the end frequencies of the fitting interval(s)}

\item{bDebug}{(TRUE) plot diagnostics}

\item{breaks}{vector of breakpoints to which the spectra is binned (optional)}

\item{indexRemove}{bins that are removed (e.g. containing the annual and semiannual cycle)}

\item{df.log}{resolution of the bins (if breaks are not provided)}

\item{i.fStart}{index of first (lowest) frequency to be used}
}
\value{
list(slope=slope,slopesd=slopesd,spec=saveSpec,freq=binFreq,intercept=intercept)
}
\author{
Thomas Laepple
}
