% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorIrregTimser.R
\name{CorIrregTimser}
\alias{CorIrregTimser}
\title{CorIrregTimser}
\usage{
CorIrregTimser(timser1, timser2, detr, method = c("InterpolationMethod",
  "DirectFiltering", "IntegrandInterpolationMethod"),
  appliedFilter = c("gauss", "runmean", "lowpass"), fc, tn = seq(from = 10,
  to = max(c(index(timser1), index(timser2))), by = 10), dt,
  int.method = c("linear", "nearest"), k = 5, filt.output)
}
\arguments{
\item{timser1, timser2}{time series (zoo-objects)}

\item{detr}{TRUE for removing a linear trend, else FALSE}

\item{method}{method to handle irregularity of sampling (DirectFiltering, IntegrandInterpolationMethod, InterpolationMethod)}

\item{appliedFilter}{time domain filter (gauss, runmean, lowpass)}

\item{fc}{cut-off frequency of the applied filter}

\item{tn}{output vector (time) of the filtered data (only used in case of DirectFiltering and IntegrandInterpolationMethod)}

\item{dt}{regular inter-observation time step of the interpolation (only used in case of InterpolationMethod)}

\item{int.method}{kind of interpolation (linear, nearest neighbor) (only used in case of InterpolationMethod)}

\item{k}{scaling factor to define the sharpness of the lowpass}

\item{filt.output}{TRUE for returning correlation ($cor) and filter results ($ft1, $ft2), FALSE for only returning correlation}
}
\description{
estimating the time scale dependent correlation of irregularly sampled time series
}
\examples{
timeseries1 <- zoo(rnorm(100), order.by=sort(runif(100,min=1,max=1000)))
timeseries2 <- zoo(rnorm(100), order.by=sort(runif(100,min=1,max=1000)))
CorIrregTimser(timser1=timeseries1, timser2=timeseries2, detr=FALSE, method="InterpolationMethod", appliedFilter="gauss", fc=1/200, tn=NA, dt=10, int.method="linear", k=NA, filt.output=TRUE)
}
